Move truth clustering to g4simulation/g4tracking node
