--- offline/packages/trackbase/TrkrHitTruthClustersLinkDef.h
@@ -1,5 +1,0 @@
-#ifdef __CINT__
-
-#pragma link C++ class TrkrHitTruthClusters+;
-
-#endif /* __CINT__ */
