--- offline/packages/trackbase/TrkrHitTruthClustersv1LinkDef.h
@@ -1,5 +1,0 @@
-#ifdef __CINT__
-
-#pragma link C++ class TrkrHitTruthClustersv1+;
-
-#endif /* __CINT__ */
