--- offline/packages/trackbase/TrkrTruthClusters.cc
@@ -1,18 +1,0 @@
-#include "TrkrTruthClusters.h"
-
-namespace
-{
-  TrkrTruthClusters::Vector dummy_vector;
-}
-
-TrkrTruthClusters::ConstRange TrkrTruthClusters::getClusters(short /*trackid*/) const
-{ return std::make_pair(dummy_vector.begin(), dummy_vector.end()); }
-
-/* TrkrTruthClusters:: */
-
-/* TrkrTruthClusters:: */
-
-/* TrkrTruthClusters:: */
-
-/* TrkrTruthClusters:: */
-
