--- offline/packages/trackbase/TrkrTruthClustersLinkDef.h
@@ -1,5 +1,0 @@
-#ifdef __CINT__
-
-#pragma link C++ class TrkrTruthClusters+;
-
-#endif /* __CINT__ */
