--- offline/packages/trackbase/TrkrTruthClustersv1LinkDef.h
@@ -1,5 +1,0 @@
-#ifdef __CINT__
-
-#pragma link C++ class TrkrTruthClustersv1+;
-
-#endif /* __CINT__ */
