+++ simulation/g4simulation/g4eval/Makefile.am
@@ -27,6 +27,7 @@ libg4eval_la_LIBADD = \
   -lg4detectors_io \
   -lg4jets_io \
   -lg4vertex_io \
+  -lg4tracking_io \
   -ltrackbase_historic_io \
   -ltrack_io \
   -ltpc_io \
