+++ simulation/g4simulation/g4tpc/readme.md
@@ -1,0 +1,6 @@
+This job run from this macro branch: 
+https://github.com/sPHENIX-Collaboration/macros/tree/QA-tracking-pythiajet
+after an automatic merging with the master branch. 
+
+The random seeds are set deterministically, which allow us to recreate each event:  
+root Fun4All_G4_sPHENIX.C\(450,\"NullInput\",\"G4sPHENIX_job_1\"\)
