+++ simulation/g4simulation/g4tracking/TrkrTruthTrackContainer.cc
@@ -14,3 +14,5 @@ TrkrTruthTrackContainer::Map& TrkrTruthTrackContainer::getMap()
 {
   return dummy_map;
 }
+
+int TrkrTruthTrackContainer::nhw_cc() { return 106; };
