+++ simulation/g4simulation/g4tracking/configure.ac
@@ -1,0 +1,23 @@
+AC_INIT(g4tracking,[1.00])
+AC_CONFIG_SRCDIR([configure.ac])
+
+AM_INIT_AUTOMAKE
+
+AC_PROG_CXX(CC g++)
+LT_INIT([disable-static])
+
+if test $ac_cv_prog_gxx = yes; then
+   CXXFLAGS="$CXXFLAGS -Wall -Wextra -Werror"
+fi
+
+case $CXX in
+ clang++)
+  CXXFLAGS="$CXXFLAGS -Wno-deprecated-copy -Wno-bitwise-instead-of-logical -Wno-unknown-warning-option"
+ ;;
+esac
+
+CINTDEFS=" -noIncludePaths  -inlineInputHeader "
+AC_SUBST(CINTDEFS)
+
+AC_CONFIG_FILES([Makefile])
+AC_OUTPUT
